# app-backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WeatherTrack",
    "version": "1.0.0",
    "description": "API for the WeatherTrack system"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/data/latest": {
      "get": {
        "summary": "Get latest data",
        "parameters": [
          {
            "in": "query",
            "name": "location",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Location"
          },
          {
            "in": "query",
            "name": "attribute",
            "schema": {
              "type": "string"
            },
            "required": false,
            "description": "Attribute name"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Data"
                }
              }
            }
          },
          "400": {
            "description": "Missing required query parameter"
          },
          "404": {
            "description": "No data found for the specified location"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/data/max": {
      "get": {
        "summary": "Get max data in timespan",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "location",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Location"
          },
          {
            "in": "query",
            "name": "attribute",
            "schema": {
              "type": "string"
            },
            "required": false,
            "description": "Attribute name"
          },
          {
            "in": "query",
            "name": "timespan",
            "schema": {
              "type": "number"
            },
            "required": false,
            "description": "Timespan in hours"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Data"
                }
              }
            }
          },
          "400": {
            "description": "Missing required query parameter"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "No data found for the specified location and timespan"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/data/avg": {
      "get": {
        "summary": "Get average data in timespan",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "location",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Location"
          },
          {
            "in": "query",
            "name": "attribute",
            "schema": {
              "type": "string"
            },
            "required": false,
            "description": "Attribute name"
          },
          {
            "in": "query",
            "name": "timespan",
            "schema": {
              "type": "number"
            },
            "required": false,
            "description": "Timespan in hours"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "average": {
                      "type": "object"
                    },
                    "timespan": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing required query parameter"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "No data found for the specified location and timespan"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/data/min": {
      "get": {
        "summary": "Get minimum data in timespan",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "location",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Location"
          },
          {
            "in": "query",
            "name": "attribute",
            "schema": {
              "type": "string"
            },
            "required": false,
            "description": "Attribute name"
          },
          {
            "in": "query",
            "name": "timespan",
            "schema": {
              "type": "number"
            },
            "required": false,
            "description": "Timespan in hours"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Data"
                }
              }
            }
          },
          "400": {
            "description": "Missing required query parameter"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "No data found for the specified location and timespan"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/data": {
      "post": {
        "summary": "Create new data",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Data"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Data"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/locations": {
      "get": {
        "summary": "Get all locations",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Location"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/data/span": {
      "get": {
        "summary": "Get data in timespan",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "location",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Location"
          },
          {
            "in": "query",
            "name": "hours",
            "schema": {
              "type": "number"
            },
            "required": true,
            "description": "Timespan in hours"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Data"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing required query parameter"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "No data found for the specified location and timespan"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/anomalies": {
      "get": {
        "summary": "Get anomalies",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Anomaly"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Anomaly": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "location": {
            "type": "string"
          },
          "attribute": {
            "type": "string"
          },
          "value": {
            "type": "number"
          }
        },
        "required": [
          "timestamp",
          "location",
          "attribute",
          "value"
        ]
      },
...
      "Location": {
        "type": "object",
        "properties": {
          "location_name": {
            "type": "string"
          },
          "sensor_identifier": {
            "type": "number"
          },
          "coordinates": {
            "type": "object",
            "properties": {
              "lat": {
                "type": "number"
              },
              "long": {
                "type": "number"
              }
            }
          }
        },
        "required": [
          "location_name",
          "sensor_identifier",
          "coordinates"
        ]
      },
...
    }
  },
...
}